backend/requirements.txt
fastapi[all]
uvicorn
pandas
numpy
yfinance
fredapi
python-dotenv
arch
aiofiles
aioredis
